LPS-192541 Extract ImportLayoutsMVCActionCommand methods to BaseExportImportMVCActionCommand and make LayoutsImportMVCActionCommand and ExportImportMVCActionCommand extends it
