+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/internal/portlet/action/ExportImportMVCActionCommand.java
@@ -21,6 +21,7 @@ import com.liferay.exportimport.kernel.model.ExportImportConfiguration;
 import com.liferay.exportimport.kernel.service.ExportImportConfigurationLocalService;
 import com.liferay.exportimport.kernel.service.ExportImportService;
 import com.liferay.exportimport.kernel.staging.Staging;
+import com.liferay.exportimport.web.internal.portlet.helper.BaseExportImportMVCActionCommand;
 import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.portal.kernel.change.tracking.CTCollectionThreadLocal;
 import com.liferay.portal.kernel.exception.LocaleException;
@@ -32,7 +33,6 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
-import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
@@ -68,7 +68,8 @@ import org.osgi.service.component.annotations.Reference;
 	},
 	service = MVCActionCommand.class
 )
-public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
+public class ExportImportMVCActionCommand
+	extends BaseExportImportMVCActionCommand {
 
 	@Override
 	protected void doProcessAction(
@@ -112,7 +113,7 @@ public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
 					group.getCtCollectionId())) {
 
 			if (cmd.equals(Constants.ADD_TEMP)) {
-				_importLayoutsMVCActionCommand.addTempFileEntry(
+				addTempFileEntry(
 					actionRequest,
 					ExportImportHelper.TEMP_FOLDER_NAME +
 						portlet.getPortletId());
@@ -125,7 +126,7 @@ public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
 				hideDefaultSuccessMessage(actionRequest);
 			}
 			else if (cmd.equals(Constants.DELETE_TEMP)) {
-				_importLayoutsMVCActionCommand.deleteTempFileEntry(
+				deleteTempFileEntry(
 					actionRequest, actionResponse,
 					ExportImportHelper.TEMP_FOLDER_NAME +
 						portlet.getPortletId());
@@ -156,7 +157,7 @@ public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
 
 				hideDefaultSuccessMessage(actionRequest);
 
-				_importLayoutsMVCActionCommand.handleUploadException(
+				handleUploadException(
 					actionRequest, actionResponse,
 					ExportImportHelper.TEMP_FOLDER_NAME +
 						portlet.getPortletId(),
@@ -245,8 +246,7 @@ public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
 
 			importData(actionRequest, inputStream);
 
-			_importLayoutsMVCActionCommand.deleteTempFileEntry(
-				groupId, folderName);
+			deleteTempFileEntry(groupId, folderName);
 		}
 	}
 
@@ -389,9 +389,6 @@ public class ExportImportMVCActionCommand extends BaseMVCActionCommand {
 
 	@Reference
 	private GroupLocalService _groupLocalService;
-
-	@Reference
-	private ImportLayoutsMVCActionCommand _importLayoutsMVCActionCommand;
 
 	@Reference
 	private Portal _portal;
